requirements.txt
Django==3.2
djangorestframework==3.12.4
Django-environ==0.4.5
psycopg2-binary>=2.9.2
Pillow==8.4.0
PyJWT==2.3.0
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
configparser==5.2.0
cryptography==36.0.0
defusedxml==0.7.1
django-allauth==0.47.0
django-phonenumber-field==6.0.0
django-phonenumbers==1.0.1
django-sendsms==0.4
idna==3.3
oauthlib==3.1.1
phonenumbers==8.12.37
pip==21.1.2
psycopg2-binary==2.9.2
pycparser==2.21
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
setuptools==57.0.0
sqlparse==0.4.2
urllib3==1.26.7
wheel==0.36.2